+++ polars/Cargo.toml
@@ -74,8 +74,8 @@ checked_arithmetic = ["polars-core/checked_arithmetic"]
 repeat_by = ["polars-core/repeat_by", "polars-lazy/repeat_by"]
 is_first = ["polars-core/is_first", "polars-lazy/is_first"]
 is_last = ["polars-core/is_last"]
-asof_join = ["polars-core/asof_join", "polars-lazy/asof_join"]
-cross_join = ["polars-core/cross_join", "polars-lazy/cross_join"]
+asof_join = ["polars-core/asof_join", "polars-lazy/asof_join", "polars-ops/asof_join"]
+cross_join = ["polars-core/cross_join", "polars-lazy/cross_join", "polars-ops/cross_join"]
 dot_product = ["polars-core/dot_product", "polars-lazy/dot_product"]
 concat_str = ["polars-core/concat_str", "polars-lazy/concat_str"]
 row_hash = ["polars-core/row_hash", "polars-lazy/row_hash"]
@@ -108,10 +108,10 @@ product = ["polars-core/product"]
 unique_counts = ["polars-core/unique_counts", "polars-lazy/unique_counts"]
 log = ["polars-ops/log", "polars-lazy/log"]
 partition_by = ["polars-core/partition_by"]
-semi_anti_join = ["polars-core/semi_anti_join", "polars-lazy/semi_anti_join"]
+semi_anti_join = ["polars-core/semi_anti_join", "polars-lazy/semi_anti_join", "polars-ops/semi_anti_join"]
 list_eval = ["polars-lazy/list_eval"]
 cumulative_eval = ["polars-lazy/cumulative_eval"]
-chunked_ids = ["polars-core/chunked_ids", "polars-lazy/chunked_ids"]
+chunked_ids = ["polars-core/chunked_ids", "polars-lazy/chunked_ids", "polars-core/chunked_ids"]
 to_dummies = ["polars-ops/to_dummies"]
 bigidx = ["polars-core/bigidx", "polars-lazy/bigidx"]
 list_to_struct = ["polars-ops/list_to_struct", "polars-lazy/list_to_struct"]
