+++ polars/polars-core/src/series/implementations/categorical.rs
@@ -10,9 +10,9 @@ use crate::chunked_array::ops::explode::ExplodeByOffsets;
 use crate::chunked_array::AsSinglePtr;
 use crate::fmt::FmtList;
 use crate::frame::groupby::*;
-#[cfg(feature = "is_in")]
-use crate::frame::hash_join::check_categorical_src;
 use crate::frame::hash_join::ZipOuterJoinColumn;
+#[cfg(feature = "is_in")]
+use crate::frame::hash_join::_check_categorical_src;
 use crate::prelude::*;
 use crate::series::implementations::SeriesWrap;
 
@@ -391,7 +391,7 @@ impl SeriesTrait for SeriesWrap<CategoricalChunked> {
 
     #[cfg(feature = "is_in")]
     fn is_in(&self, other: &Series) -> PolarsResult<BooleanChunked> {
-        check_categorical_src(self.dtype(), other.dtype())?;
+        _check_categorical_src(self.dtype(), other.dtype())?;
         self.0.logical().is_in(&other.to_physical_repr())
     }
     #[cfg(feature = "repeat_by")]
