+++ polars/polars-ops/Cargo.toml
@@ -49,3 +49,7 @@ moment = ["polars-core/moment"]
 search_sorted = []
 top_k = []
 pivot = []
+cross_join = ["polars-core/cross_join"]
+chunked_ids = ["polars-core/chunked_ids"]
+asof_join = ["polars-core/asof_join"]
+semi_anti_join = ["polars-core/semi_anti_join"]
