+++ polars/polars-ops/src/frame/mod.rs
@@ -1,5 +1,8 @@
+mod join;
 #[cfg(feature = "pivot")]
 pub mod pivot;
+
+pub use join::DataFrameJoinOps;
 #[cfg(feature = "to_dummies")]
 use polars_core::export::rayon::prelude::*;
 use polars_core::prelude::*;
