+++ polars/polars-ops/src/prelude.rs
@@ -2,5 +2,5 @@
 pub(crate) use {crate::series::*, polars_core::export::rayon::prelude::*};
 
 pub use crate::chunked_array::*;
-pub use crate::frame::DataFrameOps;
+pub use crate::frame::{DataFrameJoinOps, DataFrameOps};
 pub use crate::series::*;
