+++ polars/polars-time/src/upsample.rs
@@ -1,4 +1,5 @@
 use polars_core::prelude::*;
+use polars_ops::prelude::*;
 
 use crate::prelude::*;
