+++ polars/tests/it/joins.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "lazy")]
 use polars::prelude::*;
 
 #[test]
